feat: Finished strategy park implementation

+++ Behavioral/Strategy-Parking/AiportCalculator.cs
@@ -4,7 +4,7 @@
     {
         private const int DAILY_RATE = 50;
 
-        public int Calculate(Period period)
+        public long Calculate(Period period)
         {
             return DAILY_RATE * period.GetDiffInDays();
         }
