+++ Behavioral/Strategy-Parking/BeachCalculator.cs
@@ -3,7 +3,8 @@
     public class BeachCalculator : ITicketCalculator
     {
         private const int HOURLY_RATE = 5;
-        public int Calculate(Period period)
+
+        public long Calculate(Period period)
         {
             return HOURLY_RATE * period.GetDiffInHours();
         }
